update temperature units

+++ source/drivers/MPU6050.cpp
@@ -85,7 +85,7 @@ int MPU6050::updateSample()
         gyro.y = (((i2cData[10] << 8) | i2cData[11]));
         gyro.z = (((i2cData[12] << 8) | i2cData[13]));
 
-        int16_t t = (((i2cData[6] << 8) | i2cData[7]));
+        int t = (((i2cData[6] << 8) | i2cData[7]));
         temp = t * 10 / 34 + 3653;
 
         sample.x /= 16;
